% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnz_poisson_regression.R
\name{glm.pnz}
\alias{glm.pnz}
\title{Fit a PNZ-Poisson Generalized Linear Model using a logarithmic link function.}
\usage{
glm.pnz(formula, data, method = "L-BFGS-B", max_retries = 5)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}. The model formula specifying the response and predictors.}

\item{data}{A data frame containing the variables specified in the formula.}

\item{method}{Character string specifying the optimization method to be used in the function. Default is \code{"L-BFGS-B"}.}

\item{max_retries}{Integer specifying the maximum number of retries for the optimization algorithm.}
}
\value{
An object of class \code{pnzpois_glm} containing the fitted model results, including coefficients, fitted values, and other relevant information.
}
\description{
Fits a generalized linear model using the PNZ-Poisson distribution for the response variable. This function leverages maximum likelihood estimation with a log link function.
}
\details{
The \code{glm.pnz} function is designed to fit a PNZ-Poisson generalized linear model to the provided data. By specifying \code{method}, and \code{start_params}, users can control the optimization process used in maximum likelihood estimation. This allows for greater customization and control over the fitting process, enabling adjustments based on data characteristics or specific modeling requirements.
}
\examples{
# Fit a PNZ-Poisson model with default settings
fit <- glm.pnz(y ~ x1 + x2, data = my_data)

# Fit a PNZ-Poisson model with custom optimization method and maximum retries
fit <- glm.pnz(y ~ x1 + x2, data = my_data, method = "BFGS", max_retries = 10)

# Fit a PNZ-Poisson model with custom starting parameters
start_vals <- c(beta1 = 0.5, beta2 = -0.5, theta = 1)
fit <- glm.pnz(y ~ x1 + x2, data = my_data, start_params = start_vals)

}
