% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnz_poisson_regression.R
\name{print.summary.pnzpois_glm}
\alias{print.summary.pnzpois_glm}
\title{Print Method for Summary PNZ-Poisson GLM Objects}
\usage{
\method{print}{summary.pnzpois_glm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.pnzpois_glm} produced by \code{\link{summary.pnz_glm}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Prints the summary to the console.
}
\description{
Prints a detailed summary of a fitted PNZ-Poisson generalized linear model to the console,
including call, residual summaries, coefficients table, dispersion parameter, deviance measures,
information criteria, degrees of freedom, and number of iterations.
}
