% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnz_poisson.R
\name{rpnzpois}
\alias{rpnzpois}
\title{Random Generation from the PNZ-Poisson Distribution}
\usage{
rpnzpois(n, lambda, theta)
}
\arguments{
\item{n}{Integer specifying the number of random samples to generate.}

\item{lambda}{A positive parameter representing the mean of the distribution.}

\item{theta}{A positive parameter representing the dispersion of the distribution.}
}
\value{
An integer vector of length \code{n}, containing random samples from the PNZ-Poisson distribution.
}
\description{
Generates random samples from the PNZ-Poisson distribution.
}
\details{
The generated random sample follow the PNZ-Poisson distribution.
}
\examples{
# Generate 10 random samples with lambda = 2 and theta = 1
rpnzpois(10, lambda = 2, theta = 1)
}
